# chore(docs): speed up build (VIV-1312)