--- a/apps/docs/.eleventy.js
+++ b/apps/docs/.eleventy.js
@@ -1,6 +1,4 @@
 const { EleventyRenderPlugin } = require("@11ty/eleventy");
-const codeBlockDemo = require("./transformers/code-block-demo");
-const variablesPreview = require("./transformers/variables-preview");
 const markdownLibrary = require("./libraries/markdown-it");
 const CleanCSS = require("clean-css");
 const fs = require("fs");
@@ -37,9 +35,6 @@ module.exports = function (eleventyConfig) {
       ignorePaths: '/components/**/frames/*.html'
     }
   });
-
-  eleventyConfig.addTransform('codeBlockDemo', codeBlockDemo);
-  eleventyConfig.addTransform('variablesPreview', variablesPreview);
 
   eleventyConfig.setUseGitIgnore(false);