--- a/apps/docs/_data/icons.json
+++ b/apps/docs/_data/icons.json
@@ -1,7 +1,6 @@
 [
 	{
 		"title": "icons-gallery",
-		"markdown": "./docs/icon/icons-gallery.md",
-		"modules": []
+		"markdown": "./docs/icon/icons-gallery.md"
 	}
 ]