--- a/apps/docs/_includes/prefetches.njk
+++ b/apps/docs/_includes/prefetches.njk
@@ -1,3 +1,0 @@
-{% for component in components %}
-	<link rel="prefetch" href="{{ '/assets/modules/components/' + component.title + '/index.js' | url }}">
-{% endfor %}