--- a/apps/docs/pages/introduction.njk
+++ b/apps/docs/pages/introduction.njk
@@ -9,12 +9,6 @@ permalink: "getting-started/{{ doc.title | slug }}/"
 
 {% extends '../_layouts/base.njk' %}
 
-{% block header %}
-	{% for url in doc.modules %}
-<script type="module" src="{{ url | url }}"></script>
-	{% endfor %}
-{% endblock %}
-
 {% block content %}
 	{% renderFile doc.markdown %}
 {% endblock %}