--- a/apps/docs/transformers/code-block-demo/index.js
+++ b/apps/docs/transformers/code-block-demo/index.js
@@ -1,26 +1,17 @@
 const { JSDOM } = require('jsdom');
 const createCodeExample = require('./createCodeExample')
-const jsonData = [
-	...require('../../_data/components.json'),
-	...require('../../_data/designs.json'),
-	...require('../../_data/icons.json'),
-	...require('../../_data/introduction.json')
-];
 
 module.exports = function (content, outputPath) {
 	if (!outputPath.endsWith('.html')) {
 		return content;
 	}
 
-	const componentName = outputPath.split('/').at(-2);
-	const componentData = jsonData.find(c => c.title === componentName);
-
 	const dom = new JSDOM(content);
 	const preBlocks = dom.window.document.querySelectorAll('pre.preview');
 
 	preBlocks.forEach((pre) => {
 		const code = pre.querySelector(':scope > code');
-		pre.replaceWith(createCodeExample(code, pre, outputPath, componentData, null));
+		pre.replaceWith(createCodeExample(code, pre, outputPath, null));
 	});
 	return dom.serialize();
 };