… apps/docs/_data/components.json
… package.json